requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
PyMySQL>=1.1.0
vertica-python>=1.4.0
tabulate>=0.9.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
schedule>=1.2.0
